requirements.txt
PyQt5
mysql-connector-python
python-dotenv
bcrypt
